@@ -1,9 +1,16 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
+import { MatIconRegistry } from '@angular/material/icon';
 
 @Component({
   selector: 'app-root',
   template: '<app-navigation></app-navigation>'
 })
-export class AppComponent {
+export class AppComponent implements OnInit {
   title = 'frontend';
+
+  constructor(private matIconReg: MatIconRegistry) {}
+
+  ngOnInit() {
+    this.matIconReg.setDefaultFontSetClass('material-symbols-outlined');
+  }
 }
