@@ -11,7 +11,7 @@ import { inject } from '@angular/core';
 import { CanActivateFn } from '@angular/router';
 import { PermissionService } from 'src/app/permission.service';
 import { Event } from '../event.model';
-import { combineLatest, map } from 'rxjs';
+import { combineLatest, map, of } from 'rxjs';
 import { EventService } from '../event.service';
 
 // TODO: Refactor with a new event permission API so that we do not
@@ -38,9 +38,12 @@ export const eventEditorGuard: CanActivateFn = (route, _) => {
   );
 
   // Checks if the user is the organizer for the event
-  const isOrganizerCheck$ = inject(EventService)
-    .getEvent(+eventId)
-    .pipe(map((event) => event?.is_organizer ?? false));
+  const isOrganizerCheck$ =
+    eventId === 'new'
+      ? of(true)
+      : inject(EventService)
+          .getEvent(+eventId)
+          .pipe(map((event) => event?.is_organizer ?? false));
 
   // Since only one check has to be true for the user to see the page,
   // we combine the results of these observables into a single
