@@ -70,7 +70,6 @@ import { GroupEventsPipe } from './pipes/group-events.pipe';
     RouterModule,
     SharedModule,
     EventRoutingModule
-  ],
-  providers: [GroupEventsPipe]
+  ]
 })
 export class EventModule {}
